💄 Move CustomerLogoParser duplicate code to its own static class
